Replace Nav with Breadcrumb in navigator
Since we want a more step-through sort of graphical component for our
navigator, replace Nav with Breadcrumb and NavItem with Breadcrumb.Item.

We also need to filter the accessible components of the stepper by
progress so we don't allow citizens to select an option they have not
completed yet.  The preamble to the component list in the render()
function achieves this with the conditional in the map statement.

- Upgrade react-bootstrap to use Breadcrumb.Item instead of
  BreadcrumbItem.

Signed-off-by: Ryan Y <ryayak1460@protingumas.com>