Correct loading predicate
Change the conjunction to a disjunction in the loading predicate.
What's your function. #loldemorgan

Without a disjunction, the transition from the ward page to the survey
page would never happen due to a failure in the SurveyQuestion component
where it expected at least one question, which it would never have since
it tried to transition before it had any questions (questions.length ===
0).

Signed-off-by: Ryan Y <ryayak1460@protingumas.com>