Use "envify" to use the analytics key client-side
Use the envify-loader for webpack to convert environment variables to
plain strings at build.  The previous commits for this would not
actually use the environment variable in the react-ga components.  This,
with the global "envify" npm package (noted in the readme), will
actually convert them during build so they can exist as plain strings at
runtime.

Signed-off-by: Ryan Y <ryayak1460@protingumas.com>